feat(hazards): scrap Gust Fan; add baked-blur interior to the Vortex Well
Per playtest direction:

- Scrap the Gust Fan (too similar to a boon the operator is building). Removed
  cleanly from config, registry, draw, editor, AI (hazardRepulsion + classifier),
  mapClassifier, Codex, ARCHITECTURE, and the design doc. Vortex Well reassigned
  from id 909 -> 908 (Batch 2 is now one hazard; nothing had shipped). force-zones
  -test.js renamed/rescoped to vortex-well-test.js.

- Add a blur effect inside the Vortex Well: a hazy violet swirl baked ONCE into an
  offscreen sprite via ctx.filter blur (the getBlackoutHoleSprite pattern) and
  rotate-blitted each frame, so the blur churns without ever running a per-frame
  canvas filter (a mobile GPU killer per the cosmetic-perf-collapse history). It's
  a stylized haze, not a true frosted-glass blur of content behind it (deferred —
  that needs a per-frame framebuffer capture).

CHANGELOG/Codex/ARCHITECTURE/design-doc updated; full battery + build green.
