+++ ARCHITECTURE.md
@@ -65,7 +65,7 @@ direction is strictly downward (no cycles): `game.js` → entities → `shapes.j
 | `player.js` | ~740 | `Player` (extends `Circle`) and `LobbyStartButton`. The big behavioral class: movement, punch/puck, fire, sleep/AFK, infection, `handleHit` (type dispatcher → `handlePunchHit`/`handlePuckHit`/`handleMapCellHit`), ability pickup via the `ABILITY_TILE_CTORS` table + `tryAcquireAbility`, scoring/notches, kill/reset. |
 | `abilities.js` | ~165 | `Ability` base + `Blindfold`, `Swap`, `IceCannon`, `Bomb`, `SpeedBuff`, `SpeedDebuff`, `TileSwap`, `Cut`, `BombTrigger`. |
 | `projectiles.js` | ~145 | `Projectile` (extends `Circle`) + `CloudProj`, `SnowFlakeProj`, `BombProj`, `Puck`. |
-| `hazards.js` | ~360 | `HazardRail`, `Hazard`, `Bumper`, `BumperWall` (rotated-rect slingshot line), `Rotor` (bumper orbiting a pivot — `advance(dt)` motion hook, `streamAngle` wire), `Geyser` (timed dormant→charge→erupt vent — phase timer in `update(dt)`, `netState` wire), `Mine` (proximity-tripped one-shot blast — `netState` wire), `GustFan` (rectangular **force zone** — a steady directional push rather than a one-shot Punch; rotated Rect like `BumperWall`, `isGust`), `VortexWell` (circular force zone — a continuous inward pull with a **calm-eye** profile, `force·4r(1−r)`: zero at the dead centre and rim, peak in the mid-ring, so the centre stays escapable; `isVortex`), plus the **hazard-kind registry** (`HAZARD_KINDS`, `hazardKindById`, `registerHazardKind`) — single source of truth for map-authorable hazard kinds. **Adding a hazard kind touches:** a `config.json` `hazards` entry, one `registerHazardKind` call here (with `railed`/`directional` + `build`), a drawer in `client/scripts/draw.js` (`buildHazardDrawers`), and an editor entry in `client/scripts/create.js` (`EDITOR_HAZARD_KINDS`, with optional `paint`/`swatchPaint`/`segmentSelect` hooks for non-disc kinds). **Per-tick motion/state:** moveable kinds either ride a `rail` or define `advance(dt)` (engine.updateHazards); stationary stateful kinds run a phase timer in `update(dt)` (gameBoard.updateHazards passes dt) and ship a `netState` int. **Force-zone kinds** (`forceZone: true`) apply a continuous force via `applyForce(player)`, called ONCE per tick per player by `gameBoard.updateHazards` (NOT in `handleHit` — `engine.narrowBase` fires handleHit up to twice per overlapping pair, which would double the force; handleHit is a no-op for these), only while racing/collapsing, self-filtering by containment and skipping protected/star-power karts like `applyExplosionForce`. The AI gets an `isGust` (counter-steer the wind) / `isVortex` (route around the core) branch in `aiController` and `mapClassifier`. `gameBoard.generateHazards`, `utils.validateMap` (`directional` ⇒ finite angle), the lightning speed-up, the editor palette/swatch/thumbnail/shortcut, and the wire payloads are all registry-driven and need no changes. |
+| `hazards.js` | ~360 | `HazardRail`, `Hazard`, `Bumper`, `BumperWall` (rotated-rect slingshot line), `Rotor` (bumper orbiting a pivot — `advance(dt)` motion hook, `streamAngle` wire), `Geyser` (timed dormant→charge→erupt vent — phase timer in `update(dt)`, `netState` wire), `Mine` (proximity-tripped one-shot blast — `netState` wire), `VortexWell` (circular **force zone** — a continuous inward pull rather than a one-shot Punch, with a **calm-eye** profile `force·4r(1−r)`: zero at the dead centre and rim, peak in the mid-ring, so the centre stays escapable; `isVortex`), plus the **hazard-kind registry** (`HAZARD_KINDS`, `hazardKindById`, `registerHazardKind`) — single source of truth for map-authorable hazard kinds. **Adding a hazard kind touches:** a `config.json` `hazards` entry, one `registerHazardKind` call here (with `railed`/`directional` + `build`), a drawer in `client/scripts/draw.js` (`buildHazardDrawers`), and an editor entry in `client/scripts/create.js` (`EDITOR_HAZARD_KINDS`, with optional `paint`/`swatchPaint`/`segmentSelect` hooks for non-disc kinds). **Per-tick motion/state:** moveable kinds either ride a `rail` or define `advance(dt)` (engine.updateHazards); stationary stateful kinds run a phase timer in `update(dt)` (gameBoard.updateHazards passes dt) and ship a `netState` int. **Force-zone kinds** (`forceZone: true`) apply a continuous force via `applyForce(player)`, called ONCE per tick per player by `gameBoard.updateHazards` (NOT in `handleHit` — `engine.narrowBase` fires handleHit up to twice per overlapping pair, which would double the force; handleHit is a no-op for these), only while racing/collapsing, self-filtering by distance and skipping protected/star-power karts like `applyExplosionForce`. The AI gets an `isVortex` (route around the core) branch in `aiController` and `mapClassifier`. `gameBoard.generateHazards`, `utils.validateMap` (`directional` ⇒ finite angle), the lightning speed-up, the editor palette/swatch/thumbnail/shortcut, and the wire payloads are all registry-driven and need no changes. |
 | `boons.js` | ~70 | `Boon` (extends `Hazard`) + `DashArrows`, the **helpful** counterpart to hazards. Boons register into the SAME kind registry via `registerBoonKind` (which stamps `helpful: true` and reads ids from `config.boons`, 950+), so they live in the same runtime `hazardList` and reuse the hazard wire, drawer, validator, and editor pipeline unchanged. `helpful` is the only divergence — it gates the lightning speed-up, the AI's `hazardRepulsion`/cell-penalty (bots don't dodge a boost), and `mapClassifier`'s difficulty count. Loaded for its side-effect registrations by `hazards.js` (after that module's exports). **Adding a boon kind touches:** a `config.json` `boons` entry, one `registerBoonKind` call here (with `directional` + `build`), a drawer in `client/scripts/draw.js` (`buildHazardDrawers`), and an editor entry in `client/scripts/create.js` (`EDITOR_HAZARD_KINDS` with `group: "boon"` + a `paint`/`swatchPaint`). The map array (`hazards[]`), wire, generateHazards, validateMap, AI gate, and classifier skip are all registry/flag-driven and need no changes. |
 | `aimers.js` | ~100 | `ExplosionAimer`, `SwapAimer` (telegraph circles). |
 | `punch.js` | ~25 | `Punch` (extends `Circle`) — the melee hitbox; created by `Player` and `Bumper`. |
