+++ CHANGELOG.md
@@ -12,7 +12,7 @@ After each release the same workflow also rolls every `vX.Y.Z` from the current 
 
 ### Map editor
 
-- **Two new placeables: Gust Fan and Vortex Well.** A whole new flavour of hazard — force fields that push your kart around instead of bonking it. A **Gust Fan** blows a steady wind across a rectangular patch: set it as a crosswind over a lava bridge and it's a nerve-wracking steering test, as a tailwind and it's a committal speed lane, as a headwind and it's a toll on a shortcut. A **Vortex Well** drags everything inside its ring toward the centre — carry speed and you slingshot past, but crawl in and it sucks you down, so authors love parking one over lava or just off the racing line. Place and rotate both from the editor's Hazards palette. Rival AI racers fight the wind to hold their line and steer around the well's core.
+- **New placeable: Vortex Well.** A whole new flavour of hazard — a force field that pulls your kart around instead of bonking it. It drags everything inside its ring toward the centre: carry speed and you slingshot past, but crawl in and you're drawn down into the swirl. The very middle is calm, so if you get caught you can build up speed in the centre and power back out through the ring — annoying and slow, never a dead end (unless someone parked it over lava). Authors place it from the editor's Hazards palette; rival AI racers steer around the core.
 
 ## v0.38.2 — 2026-06-14
