+++ client/scripts/learn.js
@@ -137,7 +137,7 @@
                   detail: "A pickup pad. Roll across it to grab a random single-use power. It goes dormant briefly after someone claims it, then lights back up for the next racer." },
                 { id: "tile-bumper", name: "Bumpers", icon: art("bumper"), anim: "bumper",
                   blurb: "Springy obstacles that fling you back.",
-                  detail: "Bump one and you're flung away — maddening in a hurry, but a clever way to launch a chasing rival off their line. Some maps add moving bumpers that sweep across the track, so the safe lane keeps shifting, and some stretch the same springy sting into a bumper wall — a whole pinball-style line that kicks you straight off its face, great for shrugging karts back onto the racing line or guarding a shortcut. Other maps mount a bumper on a spinning arm — a rotor — that sweeps a full circle around its pivot like a clock hand, so the safe moment comes round and round. And watch the ground for geysers — vents that sit quiet, then bubble up a warning and erupt, launching anyone standing on or near them; the telegraph is your cue to step off. Proximity mines play the opposite game: they wait, armed and silent, until a kart strays too close, then blink down a fuse — a moment to scramble off if you're quick — before blowing up and flinging whoever's still nearby, so springing one as the leader can wreck the racers right behind you. A newer family doesn't bonk you at all — it pushes you with a steady force. Gust fans blow a constant wind across a patch of track: a crosswind shoves you sideways toward the edge so you have to lean into it, a tailwind hands you free speed, and a headwind makes a stretch a slog (and on ice, with so little grip, the wind drags you even further). Vortex wells are the opposite of a bumper: they pull everything inside their ring toward the centre, so carry enough speed and you slingshot right past, but get caught crawling and you're reeled in — which is why you'll often find one parked over lava or just off the fast line. The very centre is calm, though: if you do get pulled in, keep driving and you'll build up speed in the quiet middle and power back out through the ring — annoying and slow, but never a dead end (unless the middle was sitting on lava). Rival AI racers time the moving sweeps like you do — they hold at the edge of the bumper's path and dart through behind it, favoring the ends of its run where the gap stays open longest, they dash across a rotor's ring between passes, they steer wide of walls rather than ride them, they lean into a gust fan's wind to hold their line, and they route around a vortex well's core." },
+                  detail: "Bump one and you're flung away — maddening in a hurry, but a clever way to launch a chasing rival off their line. Some maps add moving bumpers that sweep across the track, so the safe lane keeps shifting, and some stretch the same springy sting into a bumper wall — a whole pinball-style line that kicks you straight off its face, great for shrugging karts back onto the racing line or guarding a shortcut. Other maps mount a bumper on a spinning arm — a rotor — that sweeps a full circle around its pivot like a clock hand, so the safe moment comes round and round. And watch the ground for geysers — vents that sit quiet, then bubble up a warning and erupt, launching anyone standing on or near them; the telegraph is your cue to step off. Proximity mines play the opposite game: they wait, armed and silent, until a kart strays too close, then blink down a fuse — a moment to scramble off if you're quick — before blowing up and flinging whoever's still nearby, so springing one as the leader can wreck the racers right behind you. A newer kind doesn't bonk you at all — it pulls you with a steady force. Vortex wells are the opposite of a bumper: they pull everything inside their ring toward the centre, so carry enough speed and you slingshot right past, but get caught crawling and you're reeled in — which is why you'll often find one parked over lava or just off the fast line. The very centre is calm, though: if you do get pulled in, keep driving and you'll build up speed in the quiet middle and power back out through the ring — annoying and slow, but never a dead end (unless the middle was sitting on lava). Rival AI racers time the moving sweeps like you do — they hold at the edge of the bumper's path and dart through behind it, favoring the ends of its run where the gap stays open longest, they dash across a rotor's ring between passes, they steer wide of walls rather than ride them, and they route around a vortex well's core." },
                 { id: "tile-random", name: "Random Ground", icon: swatch("#7c3aed"), anim: "randomTile",
                   blurb: "A wildcard disguised as another surface.",
                   detail: "A trickster tile. It masquerades as one of the ordinary surfaces, so you won't know whether you're about to hit speedy grass or draggy sand until you're already committed to it." }
